Rename webview package to authenticatedwebview
